refactor(cli): A.3.1 业务 inline init/studio/doctor 到 oclif Command
100% oclif-native:业务不再住独立 src/cli/commands/ directory，inline 到
src/cli/oclif/commands/*.ts。Command 拿到 typed args/flags 后直接跑业务，
不再二次 import legacy execute()。

_shared.ts utility（requireEvaluationReport / parseLastWindow / EvalResult /
ReportServer types）从 src/cli/commands/_shared.ts 上移到 src/cli/_shared.ts —
4 个 commands 共享，作为 deep utility 留在 src/cli/ 顶层（跟 i18n / cli-exit 同级）。

studio：runStudio + openWorkbench 作为 module-level helpers export from
src/cli/oclif/commands/studio.ts。studio-dev-spawn.test.ts 是 mock-based
（vi.mock node:child_process / node:os），需要 in-process import，所以 helper
保留 export。业务仍在 oclif Command file 内（不在独立 commands/ directory），
满足 100% oclif-native 的「CLI entry 全走 oclif」语义。

doctor：runDoctorCommand 业务整段 inline 到 Doctor.run() body，含
findDefaultSamplesPath / sampleSearchDirs 等 file-local helper。

assertion-language.test.ts 改 spawn dist 验证（之前 import runInit
in-process，inline 后 runInit 不存在）。

yarn lint && build && test 1494 全绿。
