+++ src/cli/_shared.ts
@@ -1,6 +1,6 @@
-import { CliExit } from '../cli-exit.js';
-import { tCli, type CliLang } from '../i18n.js';
-import type { EvaluationReport, ReportDocument } from '../../types/index.js';
+import { CliExit } from './cli-exit.js';
+import { tCli, type CliLang } from './i18n.js';
+import type { EvaluationReport, ReportDocument } from '../types/index.js';
 
 export interface EvalResult {
   report: ReportDocument;
