+++ src/cli/oclif/commands/studio.ts
@@ -1,11 +1,88 @@
+import { resolve } from 'node:path';
 import { Command, Flags } from '@oclif/core';
 import { bilingual } from '../i18n.js';
 import { runLegacyCommand } from '../run-legacy.js';
+import { tCli, type CliLang } from '../../i18n.js';
+import { DEFAULT_REPORTS_DIR } from '../../parse-run-config.js';
+import type { ReportServer } from '../../_shared.js';
+import type { StudioArgs, StudioFlags } from '../../types/cmd-flags.js';
 
-// oclif 版 studio — 透传 argv 给生产 execute()。
-// server 是长跑命令:生产 execute() 调 server.start() 后 await 返回 URL 然后 console.log
-// 但 server 自身保持 listen,Node.js 不会主动退出。oclif 路径同理:run() 不显式 return
-// 则 oclif 也不会 process.exit,server 继续 listen。Ctrl+C 走 Node 默认行为(legacy 同)。
+async function openWorkbench(url: string, lang: CliLang): Promise<void> {
+  const { execFile } = await import('node:child_process');
+  const { platform } = await import('node:os');
+  const browser = process.env.BROWSER?.trim();
+  if (browser?.toLowerCase() === 'none') return;
+
+  const os = platform();
+  const command = browser || (os === 'win32' ? 'cmd' : os === 'darwin' ? 'open' : 'xdg-open');
+  const args = os === 'win32' && !browser
+    ? ['/c', 'start', '', url]
+    : [url];
+  execFile(command, args, (err) => {
+    if (!err) return;
+    process.stderr.write(tCli('cli.studio.open_failed', lang, {
+      command,
+      message: err.message,
+    }));
+  });
+}
+
+// dev / browser-open 测试需要 mock `node:child_process` + `node:os`,通过 in-process
+// import 直接调用。把业务作为 module-level helper export(住在 oclif Command file 内,
+// 不再有独立 src/cli/commands/ directory),既保留 test 兼容又满足 100% oclif-native。
+export async function runStudio(
+  _args: StudioArgs,
+  flags: StudioFlags,
+  lang: CliLang,
+): Promise<void> {
+  const reportsDir = flags['reports-dir'] ?? DEFAULT_REPORTS_DIR;
+
+  if (flags.dev && !process.env.__OMK_DEV_CHILD) {
+    const { spawn } = await import('node:child_process');
+    const { fileURLToPath } = await import('node:url');
+    const cliPath = resolve(fileURLToPath(import.meta.url), '..', '..', '..', 'index.js');
+    const watchRoot = resolve(cliPath, '..', '..');
+    const childArgs = [
+      '--watch-path', watchRoot,
+      cliPath,
+      'studio',
+      '--port', flags.port,
+      ...(flags.host ? ['--host', flags.host] : []),
+      '--reports-dir', reportsDir,
+    ];
+    if (flags['analyses-dir']) {
+      childArgs.push('--analyses-dir', flags['analyses-dir']);
+    }
+    if (flags['observations-dir']) {
+      childArgs.push('--observations-dir', flags['observations-dir']);
+    }
+    if (flags['no-open']) {
+      childArgs.push('--no-open');
+    }
+    const child = spawn(process.execPath, childArgs, {
+      stdio: 'inherit',
+      env: { ...process.env, __OMK_DEV_CHILD: '1' },
+    });
+    child.on('exit', (code: number | null) => process.exit(code || 0));
+    return;
+  }
+
+  const { createReportServer } = await import('../../../server/report-server.js');
+  const server: ReportServer = createReportServer({
+    port: Number(flags.port),
+    ...(flags.host ? { host: flags.host } : {}),
+    reportsDir: resolve(reportsDir),
+    ...(flags['analyses-dir'] ? { analysesDir: resolve(flags['analyses-dir']) } : {}),
+    ...(flags['observations-dir'] ? { observationsDir: resolve(flags['observations-dir']) } : {}),
+  });
+
+  const url = await server.start();
+  console.log(tCli('cli.studio.started', lang, { url }));
+  console.log(tCli('cli.studio.stop_hint', lang));
+  if (!flags['no-open'] && process.stdout.isTTY) {
+    await openWorkbench(url, lang);
+  }
+}
 
 export default class Studio extends Command {
   static description = bilingual({
@@ -83,7 +160,6 @@ export default class Studio extends Command {
     const { flags } = await this.parse(Studio);
     const lang = (flags.lang ?? 'zh') as 'zh' | 'en';
     await runLegacyCommand(this, async () => {
-      const { runStudio } = await import('../../commands/studio.js');
       await runStudio({}, { ...flags, lang }, lang);
     });
   }
